+++ src/cmd/config.rs
@@ -23,6 +23,6 @@ pub fn run(sess: &Session, _matches: &ArgMatches) -> Result<()> {
         let handle = stdout.lock();
         serde_json::to_writer_pretty(handle, sess.config)
     };
-    println!("");
+    println!();
     result.map_err(|cause| Error::chain("Failed to serialize configuration.", cause))
 }